Made calculation of pseudo fitted development factors a bit clearer
On the sheet "Mack" changed the calculation of the pseudo fitted development factors so that it is clearer that they are the weighted average of the pseudo link ratios, with the weights from the original claims triangle. Deleted the rows containing the intermediate claims amounts previously used in the calculation.